refactor: change cartinfo file name to commonInfo and add timeTalbe type
